Create LibriLedger
LibriLedger is a project designed to help booksellers manage their stock! It takes a list of books with their codes (where the first letter tells you the category) and how many they have, then counts up the total number of books for each category (A, B, C, and so on). This makes it easy to see how many books are in each section of the bookstore.